requirements.txt
shiny
shinychat
chatlas
python-dotenv
faicons